refactor: move app and auth folders into resources folder
