+++ src/resources/app/app.module.ts
@@ -1,13 +1,13 @@
 import { Module, NestModule, MiddlewareConsumer, RequestMethod } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import ormconfig from '../ormconfig';
-import { BoardsModule } from '../resources/boards/boards.module';
-import { TasksModule } from '../resources/tasks/tasks.module';
-import { UsersModule } from '../resources/users/users.module';
+import ormconfig from '../../ormconfig';
+import { BoardsModule } from '../boards/boards.module';
+import { TasksModule } from '../tasks/tasks.module';
+import { UsersModule } from '../users/users.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from '../auth/auth.module';
-import { RequestLogger } from '../middleware/request-logger';
+import { RequestLogger } from '../../middleware/request-logger';
 
 @Module({
   imports: [
