+++ src/resources/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { UsersModule } from '../resources/users/users.module';
+import { UsersModule } from '../users/users.module';
 
 @Module({
   imports: [
