+++ src/resources/tasks/tasks.controller.ts
@@ -15,7 +15,7 @@ import { TasksService } from './tasks.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { Task } from './task.entity';
-import { AuthGuard } from '../../auth/auth.guard';
+import { AuthGuard } from '../auth/auth.guard';
 
 @Controller('boards/:boardId/tasks')
 @UseGuards(AuthGuard)
